+++ R/connect-utils.R
@@ -72,7 +72,7 @@ initialize_connection <- function(
   spark_context <- list(spark_context = session)
 
   # browser()
-  sc <- structure(
+  structure(
     list(
       master = master_label,
       cluster_id = cluster_id,
@@ -83,7 +83,8 @@ initialize_connection <- function(
       serverless = serverless,
       misc = misc,
       quote = quote,
-      con = structure(list(), class = c("spark_connection", "DBIConnection"))
+      con = structure(list(), class = c("spark_connection", "DBIConnection")),
+      connection_id = UUIDgenerate()
     ),
     class = c(
       con_class,
@@ -92,8 +93,6 @@ initialize_connection <- function(
       "DBIConnection"
     )
   )
-
-  sc
 }
 
 python_conn <- function(x) {