+++ R/spark-session.R
@@ -1,0 +1,21 @@
+spark_session_add_file <- function(x, sc, file_name = NULL) {
+  # Tracks in a environment variable which file names have already been
+  # uploaded to avoid doing it again. This is why the hashed values of
+  # the object in `x` is preferred, which ensure proper tracking of the
+  # actual objects that have been uploaded
+  connection_id <- sc[["connection_id"]]
+  artifacts <- pysparklyr_env[["artifacts"]][[connection_id]]
+  if (is.null(file_name)) {
+    file_name <- rlang::hash(x)
+  }
+  files <- artifacts[["files"]]
+  if (!file_name %in% artifacts[["files"]]) {
+    temp_file <- path_temp(file_name, ext = "rds")
+    saveRDS(x, temp_file)
+    sc_obj <- spark_session(sc)
+    sc_obj$addArtifact(temp_file, file = TRUE)
+    file_delete(temp_file)
+    pysparklyr_env[["artifacts"]][[connection_id]]$files <- c(file_name, files)
+  }
+  invisible()
+}