+++ tests/testthat/helper-init.R
@@ -5,6 +5,7 @@
 .test_env$started <- NULL
 .test_env$dbr <- NULL
 .test_env$databricks <- NULL
+.test_env$tune_grid <- NULL
 
 use_test_env <- function() {
   if (is.null(.test_env$env)) {