+++ NEWS.md
@@ -1,3 +1,10 @@
+# pysparklyr (dev)
+
+### Improvements
+
+- Adds support for `tune_grid_spark()`. It enables running a Tidymodels tune
+grid inside Spark Connect clusters. 
+
 # pysparklyr 0.2.0
 
 ### Fixes